backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
bcrypt==4.1.2
PyMySQL==1.1.0
